feat: rc

+++ src/main/java/org/cloud/sonic/agent/tools/file/UploadTools.java
@@ -68,7 +68,12 @@ public class UploadTools {
         ResponseEntity<JSONObject> responseEntity =
                 restTemplate.postForEntity(baseUrl + "/upload", param, JSONObject.class);
         if (responseEntity.getBody().getInteger("code") == 2000) {
-            transfer.delete();
+            if (uploadFile.exists()) {
+                uploadFile.delete();
+            }
+            if (transfer.exists()) {
+                transfer.delete();
+            }
         } else {
             logger.info("发送失败！" + responseEntity.getBody());
         }
